# Improve batching in boltdb index