This removes the timer and sync code, and replace it with a batch cursor that iterate when required.

Also improve the memory usage.

```
❯ benchstat before.txt after.txt 
name       old time/op    new time/op    delta
_Query-16    3.16µs ± 4%    1.01µs ± 5%  -67.96%  (p=0.008 n=5+5)

name       old alloc/op   new alloc/op   delta
_Query-16    1.08kB ± 1%    0.66kB ± 0%  -38.95%  (p=0.008 n=5+5)

name       old allocs/op  new allocs/op  delta
_Query-16      15.0 ± 0%      10.0 ± 0%  -33.33%  (p=0.008 n=5+5)
```